refactor(c): return -1 if readlines fail
Also, document the readlines() function.

+++ c/lib/read.h
@@ -1,3 +1,5 @@
 #pragma once
 
-int readlines(char ***lineptr, char *input);
+#include <stdio.h>
+
+ssize_t readlines(char ***lineptr, char *input);
